refactor: cleanup project entry point
